@@ -27,53 +27,57 @@
         }
     ],
     "devDependencies": {
-        "@electron-forge/cli": "6.4.2",
-        "@electron-forge/maker-deb": "6.4.2",
-        "@electron-forge/maker-dmg": "6.4.2",
-        "@electron-forge/maker-rpm": "6.4.2",
-        "@electron-forge/maker-squirrel": "6.4.2",
-        "@electron-forge/maker-zip": "6.4.2",
-        "@electron-forge/plugin-auto-unpack-natives": "6.4.2",
-        "@electron-forge/plugin-vite": "6.4.2",
-        "@electron/asar": "3.2.7",
-        "@fluentui/svg-icons": "1.1.221",
+        "@electron-forge/cli": "7.2.0",
+        "@electron-forge/core": "7.2.0",
+        "@electron-forge/maker-deb": "7.2.0",
+        "@electron-forge/maker-dmg": "7.2.0",
+        "@electron-forge/maker-rpm": "7.2.0",
+        "@electron-forge/maker-squirrel": "7.2.0",
+        "@electron-forge/maker-zip": "7.2.0",
+        "@electron-forge/plugin-auto-unpack-natives": "7.2.0",
+        "@electron-forge/plugin-vite": "7.2.0",
+        "@electron/asar": "3.2.8",
+        "@floating-ui/vue": "1.0.4",
+        "@fluentui/svg-icons": "1.1.225",
         "@fluentui/web-components": "2.5.16",
-        "@fortawesome/fontawesome-svg-core": "6.4.2",
-        "@fortawesome/pro-light-svg-icons": "6.4.2",
-        "@fortawesome/pro-solid-svg-icons": "6.4.2",
-        "@fortawesome/vue-fontawesome": "3.0.3",
         "@microsoft/fast-colors": "5.3.1",
-        "@microsoft/fast-foundation": "2.49.2",
-        "@vitejs/plugin-vue": "4.4.0",
-        "@wdio/cli": "8.18.2",
-        "@wdio/local-runner": "8.18.2",
-        "@wdio/mocha-framework": "8.18.2",
-        "@wdio/spec-reporter": "8.18.1",
-        "electron": "27.0.0",
+        "@microsoft/fast-foundation": "2.49.4",
+        "@sentry/electron": "4.17.0",
+        "@sentry/vite-plugin": "2.10.2",
+        "@vitejs/plugin-vue": "5.0.3",
+        "@wdio/cli": "8.28.6",
+        "@wdio/local-runner": "8.28.7",
+        "@wdio/mocha-framework": "8.28.6",
+        "@wdio/spec-reporter": "8.28.6",
+        "cross-env": "7.0.3",
+        "electron": "28.1.4",
         "electron-devtools-installer": "3.2.0",
-        "eslint": "8.51.0",
-        "eslint-config-prettier": "9.0.0",
-        "eslint-plugin-vue": "9.17.0",
-        "eslint-plugin-wdio": "8.8.7",
+        "eslint": "8.56.0",
+        "eslint-config-prettier": "9.1.0",
+        "eslint-plugin-prefer-arrow-functions": "3.2.4",
+        "eslint-plugin-vue": "9.20.1",
+        "eslint-plugin-wdio": "8.24.12",
         "normalize.css": "8.0.1",
-        "postcss": "8.4.31",
-        "postcss-html": "1.5.0",
-        "prettier": "3.0.3",
-        "sass": "1.69.3",
-        "stylelint": "15.10.3",
-        "stylelint-config-recommended-scss": "13.0.0",
+        "openid-client": "5.6.4",
+        "pinia": "2.1.7",
+        "postcss": "8.4.33",
+        "postcss-html": "1.6.0",
+        "prettier": "3.2.4",
+        "sass": "1.70.0",
+        "stylelint": "16.1.0",
+        "stylelint-config-recommended-scss": "14.0.0",
         "stylelint-config-recommended-vue": "1.5.0",
-        "stylelint-config-sass-guidelines": "10.0.0",
-        "stylelint-config-standard-scss": "11.0.0",
-        "vite": "4.4.11",
-        "vite-svg-loader": "4.0.0",
-        "vue": "3.3.4",
-        "vue-i18n": "9.5.0",
+        "stylelint-config-sass-guidelines": "11.0.0",
+        "stylelint-config-standard-scss": "13.0.0",
+        "vite": "5.0.12",
+        "vite-svg-loader": "5.1.0",
+        "vue": "3.4.14",
+        "vue-i18n": "9.9.0",
         "vue-router": "4.2.5",
-        "wdio-electron-service": "5.2.2"
+        "wdio-electron-service": "5.4.0"
     },
     "dependencies": {
-        "@electron-forge/core": "6.4.2",
-        "electron-squirrel-startup": "1.0.0"
+        "electron-squirrel-startup": "1.0.0",
+        "electron-store": "8.1.0"
     }
 }
