@@ -4,10 +4,7 @@
 
 <style lang="scss">
 .MainWindowPage {
-  padding: var(--base-gap);
-  position: absolute;
-  top: 0;
-  left: 0;
+  padding: var(--page-gap);
   width: 100%;
   height: 100%;
   overflow: auto;
