@@ -1,9 +1,10 @@
 <template>
   <div class="TitleBar">
-    <div class="TitleBarContent">
-      <TitleBarTitle class="TitleBar__Title" :release-tag="releaseTag" />
-      <TitleBarStatus class="TitleBar__Status" />
-      <div class="TitleBar__User">Daniel Hart</div>
+    <div class="Content">
+      <slot v-if="props.content">
+        <TitleBarTitle class="Title" :release-tag="releaseTag" />
+        <TitleBarStatus class="Status" />
+      </slot>
     </div>
   </div>
 </template>
@@ -12,14 +13,22 @@
 import { onBeforeMount, ref } from "vue";
 import TitleBarTitle from "@/components/title-bar/TitleBarTitle.vue";
 import TitleBarStatus from "@/components/title-bar/TitleBarStatus.vue";
+import TitleBarUser from "@/components/title-bar/TitleBarUser.vue";
+
+const props = defineProps({
+  content: {
+    type: Boolean,
+    default: true,
+  },
+});
 
 const releaseTag = ref(null);
 
 onBeforeMount(async () => {
-  if (await window.ukpid.isDevelopment) releaseTag.value = "Development";
-  else if (await window.ukpid.isTest) releaseTag.value = "Testing";
-  else if (await window.ukpid.isPrerelease) releaseTag.value = "Preview";
-  else if (await window.ukpid.isRelease) releaseTag.value = null;
+  if (await window.environment.isDevelopment) releaseTag.value = "Development";
+  else if (await window.environment.isTest) releaseTag.value = "Testing";
+  else if (await window.environment.isPrerelease) releaseTag.value = "Preview";
+  else if (await window.environment.isRelease) releaseTag.value = null;
 });
 </script>
 
@@ -32,33 +41,29 @@ onBeforeMount(async () => {
   position: absolute;
   top: env(titlebar-area-y);
   width: 100%;
-  user-select: none;
   -webkit-app-region: drag;
 }
 
-.TitleBarContent {
+.Content {
   width: env(titlebar-area-width);
   height: 100%;
   left: env(titlebar-area-x);
   position: absolute;
   top: 0;
   display: grid;
-  grid-template-areas: "title status user";
-  grid-template-columns: minmax(max-content, 1fr) max-content max-content;
+  grid-template-areas: "title status";
+  grid-template-columns: minmax(max-content, 1fr) max-content;
+  grid-template-rows: 100%;
   gap: fns.gap(2);
   align-content: center;
-  padding: 0 fns.gap(1);
+  padding: 0 fns.gap(1) 0 fns.gap(0.5);
 }
 
-.TitleBar__Title {
+.Title {
   grid-area: title;
 }
 
-.TitleBar__Status {
+.Status {
   grid-area: status;
-}
-
-.TitleBar__User {
-  grid-area: user;
 }
 </style>
